% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h264-interface.R
\name{h264_open}
\alias{h264_open}
\alias{h264_close}
\title{Initialise an h264 Annex B encoder}
\usage{
h264_open(con, fps = 25)

h264_close(vc)
}
\arguments{
\item{con}{Output filename. A character filename or an open binary connection}

\item{fps}{Frames per second for playback. Default: 25}

\item{vc}{An h264 \emph{video context} object. Created using \code{mp4_open()}}
}
\value{
\code{h264_open()} return a video context (\code{vc}) object. \code{h264_close()}
       does not return a value.
}
\description{
Initialise an h264 Annex B encoder
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
vc <- h264_open(tempfile())
\dontshow{\}) # examplesIf}
}
