% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp4-interface.R
\name{mp4_open}
\alias{mp4_open}
\alias{mp4_close}
\title{Initialise an mp4 encoder}
\usage{
mp4_open(con, fps = 25, mdat64 = FALSE)

mp4_close(vc)
}
\arguments{
\item{con}{Output filename. A character filename or an open binary connection 
(which is seekable)}

\item{fps}{Frmaes per second. Default: 25}

\item{mdat64}{Allow video sequences larger than ~4GB by switching to a 64-bit
length field for the core 'mdat' box.  Default: FALSE.}

\item{vc}{An mp4 \emph{video context} object. Created using \code{mp4_open()}}
}
\value{
\code{mp4_open()} return a Video Coding \code{vc} context object. 
       \code{mp4_close()}  does not return a value.
}
\description{
Initialise an mp4 encoder
}
\examples{
vc <- mp4_open(tempfile())
}
