% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp4-interface.R
\name{mp4_write}
\alias{mp4_write}
\title{Write a single frame into the mp4}
\usage{
mp4_write(vc, frame)
}
\arguments{
\item{vc}{An mp4 \emph{video context} object. Created using \code{mp4_open()}}

\item{frame}{Image data. This may be: a native raster, a 3D numeric array 
of RGB data, a 2D matrix of grey data. For array and matrix input, 
all numeric values must be in the
range [0, 1] with no NAs allowed.}
}
\value{
\code{mp4} video context (\code{vc}) is returned invisibly
}
\description{
Write a single frame into the mp4
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
vc <- mp4_open()
mp4_write(vc, frame)
mp4_close(vc)
\dontshow{\}) # examplesIf}
}
