% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp4encoder-package.R
\docType{package}
\name{mp4encoder-package}
\alias{mp4encoder}
\alias{mp4encoder-package}
\title{mp4encoder: Create Uncompressed H264 Video Streams and MP4 Video Files}
\description{
A video encoder which creates standard 'H.264' encoded video in an 'mp4' container. All encoding is done using base R, and no external encoding libraries are used - this allows the package to work on any R installation. This package only supports uncompressed video streams which are very fast to create but can be large in size.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/coolbutuseless/mp4encoder}
  \item Report bugs at \url{https://github.com/coolbutuseless/mp4encoder/issues}
}

}
\author{
\strong{Maintainer}: Mike Cheng \email{mikefc@coolbutuseless.com} [copyright holder]

}
\keyword{internal}
