% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgb-to-ycbcr.R
\name{nr_to_macroblocks_c}
\alias{nr_to_macroblocks_c}
\title{Conversion from RGB to YCbCr in C}
\usage{
nr_to_macroblocks_c(vc, nr)
}
\arguments{
\item{vc}{video context}

\item{nr}{nativeRaster}
}
\value{
list of Y Cb Cr components at 4:2:0 subsampling
}
\description{
Conversion from RGB to YCbCr in C
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
array_to_ycbcr(arr)
\dontshow{\}) # examplesIf}
}
