% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pad-to-size.R
\name{pad_array}
\alias{pad_array}
\alias{pad_nr}
\title{Pad an array or native raster so width and height are multiples of 16}
\usage{
pad_array(x, hjust = 0.5, vjust = 0.5, fill = c(0, 0, 0), dst = NULL)

pad_nr(x, hjust = 0.5, vjust = 0.5, fill = -16777216, dst = NULL)
}
\arguments{
\item{x}{array or native raster}

\item{hjust, vjust}{justification of array within expanded area. Default: (0.5, 0.5)
to center the array}

\item{fill}{background fill color. For arrays, specify a 3-element RGB numeric 
vector with values in the range [0, 1].  For native rasters, specify 
an integer color. Default: -16777216 (black).  See \code{colorfast::col_to_int()}}

\item{dst}{a pre-allocated destiation array or native raster of the correct size. 
Or NULL (the default) which will allocate a new object to hold the result.}
}
\value{
either a new 3d array or raster with the correct dimensions, or return the provided
        array if it was already correctly sized
}
\description{
Note: For maximum encoding speed, the user should
always aim to generate images with dimensions which are multiples of 16.
}
\examples{
        
arr <- array(1, c(8, 32, 3))         
arr <- pad_array(arr)
# Height is now padded to 16. Width remains unchanged (already a multiple of 16)
dim(arr)         
}
