% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgb-to-ycbcr.R
\name{rgb_to_ycbcr}
\alias{rgb_to_ycbcr}
\title{Convert RGB array to 4:2:0 YCbCr}
\usage{
rgb_to_ycbcr(im)
}
\arguments{
\item{im}{Numeric RGB array}
}
\value{
List of 3 components: Y, Cb, Cr. Each component is a numeric array at the 
        specified sub-sampling.
}
\description{
Convert RGB array to 4:2:0 YCbCr
}
\examples{
im <- array(seq(0, 1, length.out = 72), dim = c(4, 6, 3))
rgb_to_ycbcr(im)
}
