% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pad_array}
\alias{pad_array}
\title{Pad an array so width and height are a multiple of 16}
\usage{
pad_array(arr, dst = NULL)
}
\arguments{
\item{arr}{numeric 3d array}

\item{dst}{dst array of the correct size. Or NULL (the default) which will 
allocate a new array}
}
\value{
either a new 3d array with the correct dimensions, or return the provided
        array if it was already correctly sized
}
\description{
Note: Resizing an array is slow.  For maximum encoding speed, the user should
always aim to generate images with dimensions which are multiples of 16.
}
\examples{
        
arr <- array(1, c(8, 32, 3))         
arr <- pad_array(arr)
# Height is now padded to 16. Width remains unchanged (already a multiple of 16)
dim(arr)         
}
